requirements.txt
langchain
gradio
transformers
bs4
requests
torch
pillow